Number 10 fix
address parser placed westminster in the street downing st as the building and 10 as the flat.
This is because it decided that "st" is not a street (had it been "downing street" it would have been different) and this is because "st" is also short for "saint".
Added algorith match60.